requirements.txt
asgiref==3.8.1
attrs==24.3.0
bleach==6.2.0
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cssselect2==0.7.0
dicttoxml==1.7.16
Django==4.2.17
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-api-logger==1.1.16
drf-spectacular==0.28.0
fonttools==4.55.3
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
pillow==11.1.0
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
reportlab==4.2.5
requests==2.32.3
rpds-py==0.22.3
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
weasyprint==63.1
webencodings==0.5.1
zopfli==0.2.3.post1
asgiref==3.8.1
attrs==24.3.0
bleach==6.2.0
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cssselect2==0.7.0
dicttoxml==1.7.16
Django==4.2.17
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-api-logger==1.1.16
drf-spectacular==0.28.0
fonttools==4.55.3
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
pillow==11.1.0
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
reportlab==4.2.5
requests==2.32.3
rpds-py==0.22.3
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
weasyprint==63.1
webencodings==0.5.1
zopfli==0.2.3.post1
